Conversion to Maven project
