+++ MyWorld/src/main/java/com/company/Main.java
@@ -2,6 +2,8 @@ package com.company;
 
 import com.company.view.IOView;
 
+import java.sql.*;
+
 public class Main {
 
     public static void main(String[] args) {
@@ -14,5 +16,4 @@ public class Main {
         //Exit program
         System.out.println("\nFinishing myWorld, bye, bye!");
     }
-
 }
