+++ MyWorld/src/main/java/com/company/model/Ejemplar.java
