+++ MyWorld/src/main/java/com/company/model/EjemplarList.java
