+++ MyWorld/src/main/java/com/company/model/LendingMap.java
